src/test/java/com/ee/y2/robot/RobotTest.java
package com.ee.y2.robot;

import static org.junit.Assert.*;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;

@RunWith(SpringJUnit4ClassRunner.class)
@ContextConfiguration(locations = {"file:src/main/webapp/WEB-INF/spring/**/*-context.xml"})
public class RobotTest {
	
	@Autowired
	private Robot robot;
	//private Arm_Left arm_Left;
	
	
	@Test
	public void test() {

		assertNotNull(robot.getArm_Left());
		assertNotNull(robot.getArm_Right());
		
	}

}
